requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.2
ffmpeg-python==0.2.0
frozenlist==1.4.0
future==0.18.3
idna==3.6
imageio-ffmpeg==0.4.9
multidict==6.0.4
numpy==1.26.2
py-cord==2.4.1
python-dotenv==1.0.0
yarl==1.9.3
